[test] apply the new test style to databend-query - pipelines (102/177)
