+++ query/tests/it/pipelines/processors/processor_merge.rs
@@ -16,24 +16,21 @@ use std::sync::Arc;
 
 use common_base::tokio;
 use common_exception::Result;
+use databend_query::pipelines::processors::*;
 use futures::TryStreamExt;
 use pretty_assertions::assert_eq;
 
-use crate::pipelines::processors::*;
+use crate::tests;
 
 #[tokio::test(flavor = "multi_thread", worker_threads = 1)]
-async fn transform_source_test() -> Result<()> {
+async fn test_processor_merge() -> Result<()> {
     let ctx = crate::tests::create_query_context()?;
-    let test_source = crate::tests::NumberTestData::create(ctx.clone());
+    let test_source = tests::NumberTestData::create(ctx.clone());
 
-    let mut pipeline = Pipeline::create(ctx);
+    let mut pipeline = Pipeline::create(ctx.clone());
 
-    let a = test_source.number_source_transform_for_test(1)?;
-    pipeline.add_source(Arc::new(a))?;
-
-    let b = test_source.number_source_transform_for_test(1)?;
-
-    pipeline.add_source(Arc::new(b))?;
+    let source = test_source.number_source_transform_for_test(2)?;
+    pipeline.add_source(Arc::new(source))?;
     pipeline.merge_processor()?;
 
     let stream = pipeline.execute().await?;
@@ -46,7 +43,7 @@ async fn transform_source_test() -> Result<()> {
         "| number |",
         "+--------+",
         "| 0      |",
-        "| 0      |",
+        "| 1      |",
         "+--------+",
     ];
     common_datablocks::assert_blocks_sorted_eq(expected, result.as_slice());
