+++ query/tests/it/pipelines/transforms/transform_filter.rs
@@ -17,11 +17,10 @@ use std::sync::Arc;
 use common_base::tokio;
 use common_exception::Result;
 use common_planners::*;
+use databend_query::pipelines::processors::*;
+use databend_query::pipelines::transforms::*;
 use futures::TryStreamExt;
 use pretty_assertions::assert_eq;
-
-use crate::pipelines::processors::*;
-use crate::pipelines::transforms::*;
 
 #[tokio::test(flavor = "multi_thread", worker_threads = 1)]
 async fn test_transform_filter() -> Result<()> {
