+++ query/tests/it/pipelines/transforms/transform_limit.rs
@@ -17,14 +17,13 @@ use std::sync::Arc;
 use common_base::tokio;
 use common_exception::Result;
 use common_planners::*;
+use databend_query::pipelines::processors::*;
+use databend_query::pipelines::transforms::*;
 use futures::TryStreamExt;
 use pretty_assertions::assert_eq;
 
 #[tokio::test(flavor = "multi_thread", worker_threads = 1)]
 async fn test_transform_limit() -> Result<()> {
-    use crate::pipelines::processors::*;
-    use crate::pipelines::transforms::*;
-
     let testcases = vec![
         ((Some(2), 0), vec![
             "+--------+",
